Changes to main and admin screen
The menu of the main page is not centered and all the buttons have the
same size. Admin page layout and functionality (just front end) to get
answers for a given question as well as update/delete an answer for a
given question. A section to create a new question with new answers has
been created.