Changes for admin module
In the check questions section now it is possible to add more fields for
more answers (the backend was modified to send the answer rating as
well).

In the new question the data is passed to the back end: content of the
question, plus 'n' number of answers (including rating). However, at the
moment the values are not saved in the db.