Admin module, new question section
Now in the admin > new question there is check that there must be, at
least, 10 answers. Otherwise, it will throw an error. After the question
with its answers its saved a message will be displayed and the form will
be reset.

Some icons for the buttons in the admin module were changed. And a new
icon were added to refresh the list of questions (in case that a new
question has been created)